// web/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        LinkOut: {
            /** Id */
            id: string;
            /** Url */
            url: string;
            /** Title */
            title: string | null;
            /** Category */
            category: string;
            /** Summary */
            summary?: string | null;
            /** From Device */
            from_device: string | null;
            /** Session Id */
            session_id?: string | null;
            /** Created At */
            created_at: string;
        };
// ...
        SnippetOut: {
            /** Id */
            id: string;
            /** Kind */
            kind: string;
            /** Content */
            content: string | null;
            /** From Device */
            from_device: string | null;
            /** Session Id */
            session_id?: string | null;
            /** Created At */
            created_at: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}